Reimplement title and add some buttons to the toolbar.
It seems that when buttons are added, the previous trick no
longer works. So I invented my own "docked" trick.

The buttons are now in default style and icon. Custom
icon for removing completed item function is required
in the future. It's also better to design custom style
and 'more' button icon.