Make our first custom component: HeaderBar
The HeaderBar is just the titlebar or toolbar
we implemented before. Here we just make it as
a custom component in a different view class
to make our code cleaner.

One thing to be noticed: require your class before
you use the xtype. Here, we have to require
`DueList.view.HeaderBar` before using `xtype: 'headerbar'`.